fixed new bugs + code refactoring
